cleaning up the stream folder
